feat-#72: refactor doc folder
