+++ doc/otasoft-api-logo.png
