+++ doc/otasoft-core-new-architecture.png
